pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "chromadb>=1.1.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-elasticsearch>=0.3.2",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.34",
    "langchain-text-splitters>=0.3.11",
    "langchainhub>=0.1.21",
    "langgraph>=0.6.8",
    "llama-index>=0.14.3",
    "llama-index-embeddings-gemini>=0.4.1",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-utils-workflow>=0.4.1",
    "llama-index-vector-stores-chroma>=0.5.3",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]
